Move code back to emit.rs
