+++ cli/emit.rs
@@ -4,6 +4,8 @@
 //! populate a cache, emit files, and transform a graph into the structures for
 //! loading into an isolate.
 
+use crate::args::config_file::IgnoredCompilerOptions;
+use crate::args::ConfigFile;
 use crate::args::EmitConfigOptions;
 use crate::args::TsConfig;
 use crate::args::TypeCheckMode;
@@ -26,6 +28,7 @@ use deno_core::serde::Deserializer;
 use deno_core::serde::Serialize;
 use deno_core::serde::Serializer;
 use deno_core::serde_json;
+use deno_core::serde_json::json;
 use deno_core::ModuleSpecifier;
 use deno_graph::MediaType;
 use deno_graph::ModuleGraph;
@@ -155,6 +158,136 @@ impl fmt::Display for Stats {
 
     Ok(())
   }
+}
+
+/// Represents the "default" type library that should be used when type
+/// checking the code in the module graph.  Note that a user provided config
+/// of `"lib"` would override this value.
+#[derive(Debug, Clone, Copy, Eq, Hash, PartialEq)]
+pub enum TsTypeLib {
+  DenoWindow,
+  DenoWorker,
+  UnstableDenoWindow,
+  UnstableDenoWorker,
+}
+
+impl Default for TsTypeLib {
+  fn default() -> Self {
+    Self::DenoWindow
+  }
+}
+
+impl Serialize for TsTypeLib {
+  fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+  where
+    S: Serializer,
+  {
+    let value = match self {
+      Self::DenoWindow => vec!["deno.window".to_string()],
+      Self::DenoWorker => vec!["deno.worker".to_string()],
+      Self::UnstableDenoWindow => {
+        vec!["deno.window".to_string(), "deno.unstable".to_string()]
+      }
+      Self::UnstableDenoWorker => {
+        vec!["deno.worker".to_string(), "deno.unstable".to_string()]
+      }
+    };
+    Serialize::serialize(&value, serializer)
+  }
+}
+
+/// An enum that represents the base tsc configuration to return.
+pub enum TsConfigType {
+  /// Return a configuration for bundling, using swc to emit the bundle. This is
+  /// independent of type checking.
+  Bundle,
+  /// Return a configuration to use tsc to type check and optionally emit. This
+  /// is independent of either bundling or just emitting via swc
+  Check { lib: TsTypeLib, tsc_emit: bool },
+  /// Return a configuration to use swc to emit single module files.
+  Emit,
+}
+
+pub struct TsConfigWithIgnoredOptions {
+  pub ts_config: TsConfig,
+  pub maybe_ignored_options: Option<IgnoredCompilerOptions>,
+}
+
+/// For a given configuration type and optionally a configuration file,
+/// return a `TsConfig` struct and optionally any user configuration
+/// options that were ignored.
+pub fn get_ts_config_for_emit(
+  config_type: TsConfigType,
+  maybe_config_file: Option<&ConfigFile>,
+) -> Result<TsConfigWithIgnoredOptions, AnyError> {
+  let mut ts_config = match config_type {
+    TsConfigType::Bundle => TsConfig::new(json!({
+      "checkJs": false,
+      "emitDecoratorMetadata": false,
+      "importsNotUsedAsValues": "remove",
+      "inlineSourceMap": false,
+      "inlineSources": false,
+      "sourceMap": false,
+      "jsx": "react",
+      "jsxFactory": "React.createElement",
+      "jsxFragmentFactory": "React.Fragment",
+    })),
+    TsConfigType::Check { tsc_emit, lib } => {
+      let mut ts_config = TsConfig::new(json!({
+        "allowJs": true,
+        "allowSyntheticDefaultImports": true,
+        "experimentalDecorators": true,
+        "incremental": true,
+        "jsx": "react",
+        "isolatedModules": true,
+        "lib": lib,
+        "module": "esnext",
+        "resolveJsonModule": true,
+        "strict": true,
+        "target": "esnext",
+        "tsBuildInfoFile": "deno:///.tsbuildinfo",
+        "useDefineForClassFields": true,
+        // TODO(@kitsonk) remove for Deno 2.0
+        "useUnknownInCatchVariables": false,
+      }));
+      if tsc_emit {
+        ts_config.merge(&json!({
+          "emitDecoratorMetadata": false,
+          "importsNotUsedAsValues": "remove",
+          "inlineSourceMap": true,
+          "inlineSources": true,
+          "outDir": "deno://",
+          "removeComments": true,
+        }));
+      } else {
+        ts_config.merge(&json!({
+          "noEmit": true,
+        }));
+      }
+      ts_config
+    }
+    TsConfigType::Emit => TsConfig::new(json!({
+      "checkJs": false,
+      "emitDecoratorMetadata": false,
+      "importsNotUsedAsValues": "remove",
+      "inlineSourceMap": true,
+      "inlineSources": true,
+      "sourceMap": false,
+      "jsx": "react",
+      "jsxFactory": "React.createElement",
+      "jsxFragmentFactory": "React.Fragment",
+      "resolveJsonModule": true,
+    })),
+  };
+  let maybe_ignored_options =
+    ts_config.merge_tsconfig_from_config_file(maybe_config_file)?;
+  ts_config.merge(&json!({
+    "moduleDetection": "force",
+  }));
+  Ok(TsConfigWithIgnoredOptions {
+    ts_config,
+    maybe_ignored_options,
+  })
 }
 
 /// Transform the graph into root specifiers that we can feed `tsc`. We have to
